Migrate all entity ID types from int to long
Updated all primary and foreign key types from int to long across domain entities, repositories, DTOs, services, and related interfaces. Adjusted method signatures, constructors, and property types to ensure consistency and support a larger range of unique identifiers.

+++ src/Core/Restaurant.Application/Features/Schedule/Reservations/Commands/Create/CreateReservationSpecification.cs
@@ -15,7 +15,7 @@ namespace Restaurant.Application.Features.Schedule.Reservations.Commands.Create
                                         .ThenInclude(rt => rt.RestaurantTable));
         }
 
-        public void ApplyCriteria(int id)
+        public void ApplyCriteria(long id)
         {
             AddCriteria(x => x.Id == id);
         }
