+++ src/Core/Restaurant.Domain/Entities/Production/RecipeIngredient.cs
@@ -6,12 +6,12 @@ namespace Restaurant.Domain.Entities.Production
 {
     public partial class RecipeIngredient : SoftDeletableEntity
     {
-        public int RecipeId { get; private set; }
-        public int IngredientId { get; private set; }
+        public long RecipeId { get; private set; }
+        public long IngredientId { get; private set; }
 
         public decimal Quantity { get; private set; }
 
-        public int UnitId { get; private set; }
+        public long UnitId { get; private set; }
 
         public Recipe Recipe { get; private set; } = null!;
         public Ingredient Ingredient { get; private set; } = null!;
@@ -20,19 +20,19 @@ namespace Restaurant.Domain.Entities.Production
 
     public partial class RecipeIngredient
     {
-        public RecipeIngredient SetIngredient(int ingredientId)
+        public RecipeIngredient SetIngredient(long ingredientId)
         {
             IngredientId = ingredientId;
             return this;
         }
 
-        public RecipeIngredient SetRecipe(int recipeId)
+        public RecipeIngredient SetRecipe(long recipeId)
         {
             RecipeId = recipeId;
             return this;
         }
 
-        public RecipeIngredient SetUnit(int unitId)
+        public RecipeIngredient SetUnit(long unitId)
         {
             UnitId = unitId;
             return this;
