+++ src/Core/Restaurant.Domain/Entities/Territory/RestaurantTable.cs
@@ -6,7 +6,7 @@ namespace Restaurant.Domain.Entities.Territory
 {
     public partial class RestaurantTable : SoftDeletableEntity
     {
-        public int AreaId { get; private set; }
+        public long AreaId { get; private set; }
 
         public string TableNumber { get; private set; } = null!;
         public int Capacity { get; private set; }
@@ -31,7 +31,7 @@ namespace Restaurant.Domain.Entities.Territory
     {
         public RestaurantTable() { }
 
-        public RestaurantTable SetArea(int areaId)
+        public RestaurantTable SetArea(long areaId)
         {
             AreaId = areaId;
             return this;
