+++ src/Core/Restaurant.Domain/Repositories/Catalog/IIngredientCategoryRepository.cs
@@ -4,7 +4,7 @@ namespace Restaurant.Domain.Repositories.Catalog
 {
     public interface IIngredientCategoryRepository : IRepository<IngredientCategory>
     {
-        Task<IngredientCategory?> FindByIdAsync(int id, CancellationToken cancellationToken = default);
+        Task<IngredientCategory?> FindByIdAsync(long id, CancellationToken cancellationToken = default);
         Task<IngredientCategory?> FindByIdAsync(Guid id, CancellationToken cancellationToken = default);
         Task<IngredientCategory?> FindByNameAsync(string name, CancellationToken cancellationToken = default);
