+++ src/Core/Restaurant.Domain/Repositories/Identity/IRoleRepository.cs
@@ -4,7 +4,7 @@ namespace Restaurant.Domain.Repositories.Identity
 {
     public interface IRoleRepository : IRepository<Role>
     {
-        Task<Role?> FindByIdAsync(int id, CancellationToken cancellationToken = default);
+        Task<Role?> FindByIdAsync(long id, CancellationToken cancellationToken = default);
         Task<Role?> FindByIdAsync(Guid id, CancellationToken cancellationToken = default);
 
         Task<Role?> FindByNameAsync(string name, CancellationToken cancellationToken = default);
