+++ src/Core/Restaurant.Domain/Repositories/Personnel/IDepartmentRepository.cs
@@ -4,7 +4,7 @@ namespace Restaurant.Domain.Repositories.Personnel
 {
     public interface IDepartmentRepository : IRepository<Department>
     {
-        Task<Department?> FindByIdAsync(int id, CancellationToken cancellationToken = default);
+        Task<Department?> FindByIdAsync(long id, CancellationToken cancellationToken = default);
         Task<Department?> FindByIdAsync(Guid id, CancellationToken cancellationToken = default);
         Task<Department?> FindByNameAsync(string name, CancellationToken cancellationToken = default);
