+++ src/Core/Restaurant.Domain/Repositories/Territory/IAreaRepository.cs
@@ -4,8 +4,8 @@ namespace Restaurant.Domain.Repositories.Territory
 {
     public interface IAreaRepository : IRepository<Area>
     {
-        Task<Area?> FindByIdAsync(int id, CancellationToken cancellationToken = default);
+        Task<Area?> FindByIdAsync(long id, CancellationToken cancellationToken = default);
         Task<Area?> FindByIdAsync(Guid id, CancellationToken cancellationToken = default);
-        Task<bool> IsExistingNameAsync(int branchId, string name, CancellationToken cancellationToken = default);
+        Task<bool> IsExistingNameAsync(long branchId, string name, CancellationToken cancellationToken = default);
     }
 }
